day 6 task, refactor of day 1 and 2

+++ AdventOfCode2021/Day3.h
@@ -5,7 +5,7 @@
 long long BinaryDiagnostic1(const std::string &inputpath);
 long long BinaryDiagnostic2(const std::string &inputpath);
 
-std::vector<int> InitializeArrayInt(int size);
+std::vector<int> InitializeArrayInt(size_t size);
 size_t GetCountOfBinaryData(const std::string &inputpath, std::vector<int> &zeroes, std::vector<int> &ones);
 
 char GetCommonBitAtPosition(bool max, int pos, const std::vector<std::string> &input);
